+++ benchmarks/lib/semaphore.js
@@ -1,5 +1,6 @@
 var Semaphore = module.exports = function(keyName, subscribeRedis, publishRedis) {
     this.subscribeRedis = subscribeRedis;
+    this.subscribeRedis.on('message', this._decrement.bind(this));
     this.publishRedis = publishRedis;
 
     this.keyName = keyName;
@@ -13,7 +14,7 @@ var Semaphore = module.exports = function(keyName, subscribeRedis, publishRedis)
 Semaphore.prototype.start = function(count) {
     this._count = count;
     this._isReached = false;
-    this.subscribeRedis.subscribeTo(this.keyName, this._decrement.bind(this));
+    this.subscribeRedis.subscribe(this.keyName);
 };
 
 Semaphore.prototype.reach = function(callback) {
@@ -31,7 +32,7 @@ Semaphore.prototype._decrement = function(channel, message) {
     this._count--;
 
     if (this._count === 0) {
-        this.subscribeRedis.unsubscribeFrom(this.keyName);
+        this.subscribeRedis.unsubscribe(this.keyName);
         this._callback();
     }
 };